+++ documentation/ChangeLog.md
@@ -25,6 +25,7 @@
  * DUB builder: incremental builds are now retried if previous 'dub build' exited with non-zero return code.
  * Fixed exception when dub.json modified and Project Explorer view was closed.
  * Minor improvements and bugfixes to DUB elements in Project Explorer.
+ * Added to Project Explorer context menu some DUB actions to add/remove project from local packages. 
  TODO:
    * DLTK Model elements in Project Explorer
      * flatten root script folder
@@ -32,7 +33,6 @@
      * review Script label provider (text+image) 
    * better support for sub-packages
    * support for configurations
-   * some helpful actions
 
 ### DDT 0.9.1 (2014-02-07)
  * Added debug hover to D editor. When a CDT debug session is active, this hover presents detailed info for the variable under the cursor, similar to the Variables view.